Change the VGA driver
Keep the  hsync and vsync module (reserve it in case to go back) but remove them from the game module.
Rewrite the VGA driver to make it clearer and easier to read. The new module are supposed to replace the hsync and vsync module and generate the similar output signal.